requirements.txt
Django==3.2.19
channels==3.0.4
channels-redis==3.4.1
django-allauth==0.54.0
Pillow==9.5.0
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
whitenoise==6.5.0
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.6
uvicorn[standard]==0.23.2
python-decouple==3.8
daphne==3.0.2
